+++ example/lib/env1.dart
@@ -1,22 +1,23 @@
 import 'package:envystic/envystic.dart';
 
-part 'env.g.dart';
+part 'env1.g.dart';
 
 @Envystic(path: '.env.example', encryptionKey: 'EncryptMorePlease')
-abstract class Env {
-  const factory Env() = _$Env;
+class Env1 with _$Env1 {
+  const Env1();
 
-  const Env._();
-
-  @envysticField // Default key name assigned: 'KEY1'
+  @override
+  @envysticField // Default env key name assigned: 'KEY1'
   String get key1;
 
+  @override
   @EnvysticField(name: 'FOO') // The value from 'FOO' in .env will be used
   int? get key2;
 
+  @override
   @EnvysticField(
-      name: 'MY_SPECIAL_KEY') // Test from System environment variables
-  int? get specialKey;
+      name: 'MY_SPECIAL_KEY') // Pulled from system environment variables
+  int get specialKey;
 
   // ignored
   String get drink => 'Coffee';