+++ packages/envystic/lib/envystic.dart
@@ -1,6 +1,7 @@
 library envystic;
 
 export 'src/annotation/envystic.dart';
+export 'src/annotation/envystic_all.dart';
 export 'src/annotation/envystic_field.dart';
 export 'src/annotation/key_format.dart';
 export 'src/envystic_interface.dart';