+++ packages/envystic_generator/lib/src/helpers.dart
@@ -3,87 +3,24 @@ import 'package:analyzer/dart/element/element.dart';
 import 'environment_field.dart';
 import 'fields.dart';
 
-/// Returns a quoted String literal for [value] that can be used in generated
-/// Dart code.
-String escapeDartString(String value) {
-  var hasSingleQuote = false;
-  var hasDoubleQuote = false;
-  var hasDollar = false;
-  var canBeRaw = true;
+/// An extension method for the `FieldElement` class.
+extension FieldElementExt on FieldElement {
+  /// Retrieves the type prefix of the `FieldElement`.
+  ///
+  /// This method returns the prefix of the library that defines the `FieldElement`
+  /// type, if it is imported in the current library where this extension is used.
+  /// If the type is not imported, it returns `null`.
+  String? get typePrefix {
+    final identifier = type.element?.library?.identifier;
+    if (identifier == null) return null;
 
-  value = value.replaceAllMapped(_escapeRegExp, (match) {
-    final value = match[0]!;
-    if (value == "'") {
-      hasSingleQuote = true;
-      return value;
-    } else if (value == '"') {
-      hasDoubleQuote = true;
-      return value;
-    } else if (value == r'$') {
-      hasDollar = true;
-      return value;
+    for (final e in library.libraryImports) {
+      if (e.importedLibrary?.identifier != identifier) continue;
+      return e.prefix?.element.name;
     }
-
-    canBeRaw = false;
-    return _escapeMap[value] ?? _getHexLiteral(value);
-  });
-
-  if (!hasDollar) {
-    if (hasSingleQuote) {
-      if (!hasDoubleQuote) {
-        return '"$value"';
-      }
-      // something
-    } else {
-      // trivial!
-      return "'$value'";
-    }
+    return null;
   }
-
-  if (hasDollar && canBeRaw) {
-    if (hasSingleQuote) {
-      if (!hasDoubleQuote) {
-        // quote it with single quotes!
-        return 'r"$value"';
-      }
-    } else {
-      // quote it with single quotes!
-      return "r'$value'";
-    }
-  }
-
-  // The only safe way to wrap the content is to escape all of the
-  // problematic characters - `$`, `'`, and `"`
-  final string = value.replaceAll(_dollarQuoteRegexp, r'\');
-  return "'$string'";
 }
-
-/// A [RegExp] that matches whitespace characters that should be escaped and
-/// single-quote, double-quote, and `$`
-final _escapeRegExp = RegExp('[\$\'"\\x00-\\x07\\x0E-\\x1F$_escapeMapRegexp]');
-
-final _escapeMapRegexp = _escapeMap.keys.map(_getHexLiteral).join();
-
-/// A [Map] between whitespace characters & `\` and their escape sequences.
-const _escapeMap = {
-  '\b': r'\b', // 08 - backspace
-  '\t': r'\t', // 09 - tab
-  '\n': r'\n', // 0A - new line
-  '\v': r'\v', // 0B - vertical tab
-  '\f': r'\f', // 0C - form feed
-  '\r': r'\r', // 0D - carriage return
-  '\x7F': r'\x7F', // delete
-  r'\': r'\\' // backslash
-};
-
-/// Given single-character string, return the hex-escaped equivalent.
-String _getHexLiteral(String input) {
-  final rune = input.runes.single;
-  final value = rune.toRadixString(16).toUpperCase().padLeft(2, '0');
-  return '\\x$value';
-}
-
-final _dollarQuoteRegexp = RegExp(r"""(?=[$'])""");
 
 InterfaceElement getInterface(Element element) {
   //assert(
@@ -158,4 +95,20 @@ List<EnvironmentField> getAccessors(
   }
 
   return [...getters, ...setters];
+}
+
+String incrementLastInt(String input) {
+  if (input.trim().isEmpty) return input;
+  RegExp regex = RegExp(
+      r'\d+$'); // Regular expression to find the last integer in the string
+  Match? match = regex.firstMatch(input);
+
+  if (match != null) {
+    int lastInt = int.parse(match.group(0)!);
+    String prefix = input.substring(0, match.start);
+    String newInt = (lastInt + 1).toString();
+    return prefix + newInt;
+  } else {
+    return '${input}2';
+  }
 }